+++ templates/ice-react-material-template/template/scaffold/src/menuConfig.js
@@ -1,0 +1,9 @@
+// 菜单配置
+// headerMenuConfig：头部导航配置
+// asideMenuConfig：侧边导航配置
+
+// const headerMenuConfig = [];
+
+// const asideMenuConfig = [];
+
+// export { headerMenuConfig, asideMenuConfig };