+++ templates/ice-vue-material-template/template/block/src/index.vue
@@ -1,0 +1,29 @@
+<template>
+  <div class="example-block">
+    <basic-container>
+      <h1>ExampleBlock</h1>
+    </basic-container>
+  </div>
+</template>
+
+<script>
+
+export default {
+  name: 'ExampleBlock',
+
+  data() {
+    return {}
+  },
+
+  created() {},
+
+  methods: {}
+}
+
+</script>
+
+<style>
+  .example-block {
+
+  }
+</style>