+++ tools/webpack-plugin-import/lib/webpack-plugin-import.js
@@ -25,17 +25,19 @@ module.exports = class WebpackPluginImport {
     }
   }
 
-  // 在 package.json 中通过 componentConfig 字段标记是否为 ICE 组件，如果是，则自动引入样式
-  componentCheck(result) {
+  getStylePath(result) {
     if (
       result &&
       result.resourceResolveData &&
       result.resourceResolveData.descriptionFileData &&
-      result.resourceResolveData.descriptionFileData.name ===
-        result.rawRequest &&
-      result.resourceResolveData.descriptionFileData.componentConfig
+      result.resourceResolveData.descriptionFileData.name === result.rawRequest
     ) {
-      return true;
+      // 存量 ice 组件通过 componentConfig 字段标记是否为组件，如果是 ICE 组件，则需要引入样式
+      if (result.resourceResolveData.descriptionFileData.componentConfig) {
+        return 'style.js';
+      }
+      // 在 package.json 中通过 stylePath 字段标记是否需要自动引入样式，如果配置 style path，则自动引入对应样式
+      return result.resourceResolveData.descriptionFileData.stylePath;
     }
   }
 
@@ -75,7 +77,11 @@ module.exports = class WebpackPluginImport {
                 }
 
                 if (!needAdditionalStyle) {
-                  needAdditionalStyle = this.componentCheck(result);
+                  const customStylePath = this.getStylePath(result);
+                  if (customStylePath) {
+                    stylePath = customStylePath;
+                    needAdditionalStyle = true;
+                  }
                 }
 
                 if (needAdditionalStyle) {